+++ Src/Security/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace FastEndpoints.Security;
-
-[HideFromDocs]
-public static class Constants
-{
-    //must match with FastEnpoints.Constants.PermissionsClaimType
-    public const string PermissionsClaimType = "permissions";
-}