+++ Src/Security/CookieAuth.cs
@@ -35,7 +35,7 @@ public static class CookieAuth
             claimList.AddRange(privs.Claims);
 
         if (privs.Permissions.Count > 0)
-            claimList.AddRange(privs.Permissions.Select(p => new Claim(Constants.PermissionsClaimType, p)));
+            claimList.AddRange(privs.Permissions.Select(p => new Claim(Config.SecOpts.PermissionsClaimType, p)));
 
         if (privs.Roles.Count > 0)
             claimList.AddRange(privs.Roles.Select(r => new Claim(ClaimTypes.Role, r)));