+++ Src/Library/changelog.md
@@ -1,2 +1,3 @@
 ### IMPROVEMENTS
 - swagger support for routes params with mixed patterns #380
+- misc. improvements to `FastEndpoints.Security` package internals